Implement meta-data acquisition functionality
We first download the meta-data catalog from Project Gutenberg. This catalog is
a collection of RDF+XML files, one per text, that describe various attributes of
the texts.

We then build one big RDF graph out of all of these individual descriptions and
store that graph on disk locally. It takes a while to construct this graph (this
should probably be done in a daily cron job to keep the meta-data up to date
with the catalog) but using one big graph over numerous small graphs has a
number of advantages including being able to run meta-data queries like that
return all the texts that have a certain meta-data value.